requirements.txt
absl-py==1.4.0
accelerate==0.28.0
aif360==0.6.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alphashape==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
BlackBoxAuditing==0.1.54
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.1
decorator==5.1.1
diffprivlib==0.6.4
dill==0.3.8
distro==1.9.0
dm-tree==0.1.8
ete3==3.1.3
etils==1.8.0
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.19.1
ffmpy==0.4.0
filelock==3.13.1
fiona==1.9.6
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
gensim==4.3.2
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googleapis-common-protos==1.63.0
gradio_client==1.3.0
groq==0.11.0
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.1
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mysql-connector==2.2.9
mysql-connector-python==8.3.0
namex==0.0.8
narwhals==1.8.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.2
openai==1.46.0
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.0
pandas==2.1.4
parso==0.8.3
pdfminer==20191125
pdfminer.six==20231228
peft @ git+https://github.com/huggingface/peft.git@6dca6d22922fcb1ead828b5b3c146911d7b693fb
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.21.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
PyAutoGUI==0.9.54
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.8.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.9.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
Rtree==1.2.0
rubicon-objc==0.4.7
ruff==0.6.5
safetensors==0.4.2
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
selenium==4.16.0
semantic-version==2.10.0
shap==0.45.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
slicer==0.0.7
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.38.5
streamlit==1.38.0
sympy==1.12
tabula-py==2.9.0
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-datasets==4.9.4
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.4.18
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
torch==2.3.0
torchaudio==2.2.2
torchtext==0.18.0
torchvision==0.16.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers @ git+https://github.com/huggingface/transformers.git@b382a09e28c7e59129246ccdf4b00f2cac4547a4
trimesh==4.4.0
trio==0.23.1
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.10.0
tzdata==2023.3
udocker==1.3.16
unpywall==0.2.2
urllib3==2.1.0
uvicorn==0.30.6
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
